+++ apollo-configservice/src/main/java/com/ctrip/framework/apollo/configservice/filter/ClientAuthenticationFilter.java
@@ -17,9 +17,11 @@
 package com.ctrip.framework.apollo.configservice.filter;
 
 import com.ctrip.framework.apollo.biz.config.BizConfig;
+import com.ctrip.framework.apollo.common.utils.WebUtils;
 import com.ctrip.framework.apollo.configservice.util.AccessKeyUtil;
 import com.ctrip.framework.apollo.core.signature.Signature;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
+import com.ctrip.framework.apollo.tracer.Tracer;
 import com.google.common.net.HttpHeaders;
 import java.io.IOException;
 import java.util.List;
@@ -70,27 +72,58 @@ public class ClientAuthenticationFilter implements Filter {
 
     List<String> availableSecrets = accessKeyUtil.findAvailableSecret(appId);
     if (!CollectionUtils.isEmpty(availableSecrets)) {
-      String timestamp = request.getHeader(Signature.HTTP_HEADER_TIMESTAMP);
-      String authorization = request.getHeader(HttpHeaders.AUTHORIZATION);
-
-      // check timestamp, valid within 1 minute
-      if (!checkTimestamp(timestamp)) {
-        logger.warn("Invalid timestamp. appId={},timestamp={}", appId, timestamp);
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "RequestTimeTooSkewed");
+      if (!doCheck(request, response, appId, availableSecrets, false)) {
         return;
       }
-
-      // check signature
-      String uri = request.getRequestURI();
-      String query = request.getQueryString();
-      if (!checkAuthorization(authorization, availableSecrets, timestamp, uri, query)) {
-        logger.warn("Invalid authorization. appId={},authorization={}", appId, authorization);
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
-        return;
+    } else {
+      // pre-check for observable secrets
+      List<String> observableSecrets = accessKeyUtil.findObservableSecrets(appId);
+      if (!CollectionUtils.isEmpty(observableSecrets)) {
+        doCheck(request, response, appId, observableSecrets, true);
       }
     }
 
     chain.doFilter(request, response);
+  }
+
+  /**
+   * Performs authentication checks(timestamp and signature) for the request.
+   *
+   * @param preCheck Boolean flag indicating whether this is a pre-check
+   * @return true if authentication checks is successful, false otherwise
+   */
+  private boolean doCheck(HttpServletRequest req, HttpServletResponse resp,
+      String appId, List<String> secrets, boolean preCheck) throws IOException {
+
+    String timestamp = req.getHeader(Signature.HTTP_HEADER_TIMESTAMP);
+    String authorization = req.getHeader(HttpHeaders.AUTHORIZATION);
+    String ip = WebUtils.tryToGetClientIp(req);
+
+    // check timestamp, valid within 1 minute
+    if (!checkTimestamp(timestamp)) {
+      if (preCheck) {
+        preCheckInvalidLogging(String.format("Invalid timestamp in pre-check. "
+            + "appId=%s,clientIp=%s,timestamp=%s", appId, ip, timestamp));
+      } else {
+        logger.warn("Invalid timestamp. appId={},clientIp={},timestamp={}", appId, ip, timestamp);
+        resp.sendError(HttpServletResponse.SC_UNAUTHORIZED, "RequestTimeTooSkewed");
+        return false;
+      }
+    }
+
+    // check signature
+    if (!checkAuthorization(authorization, secrets, timestamp, req.getRequestURI(), req.getQueryString())) {
+      if (preCheck) {
+        preCheckInvalidLogging(String.format("Invalid authorization in pre-check. "
+            + "appId=%s,clientIp=%s,authorization=%s", appId, ip, authorization));
+      } else {
+        logger.warn("Invalid authorization. appId={},clientIp={},authorization={}", appId, ip, authorization);
+        resp.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
+        return false;
+      }
+    }
+
+    return true;
   }
 
   @Override
@@ -129,5 +162,10 @@ public class ClientAuthenticationFilter implements Filter {
       }
     }
     return false;
+  }
+
+  protected void preCheckInvalidLogging(String message) {
+    logger.warn(message);
+    Tracer.logEvent("Apollo.AccessKey.PreCheck", message);
   }
 }